Add pom.xml to be used with Maven builds
Maven is a pretty standard project management tool, and having to copy
the dependencies by hand is kind of a pain. By configuring your IDE to
pull dependencies from the Maven repo, it's always up to date. If at
any point, you decide to use a different version of some dependency,
all you must do is update the pom.xml and anyone who pulls the updated
code will automatically have the dependencies updated for them.

This also changes the README file to README.md, so that Github can parse
the data as a Markdown file, and display it properly on the project
page.